Connecting front end views with back end parsing
